backend/requirements.txt
asgiref==3.7.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.6
pytz==2023.3
requests==2.28.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13